+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Caldav: Sort personal calendar on top as Sabredav picks the first in the list for invites.
 - Projects: Fixed issue where files and comments wouldn't display on project
 - Newsletters: fix error in smtp workflow
+- Files: make trash sortable by name, path, deletion date, deletedBy user
 
 02-09-2025: 25.0.50
 - ActiveSync: Fixed Error in modules/z-push/backend/go/TaskStore.php at line 237: Typed property